Foolishness! Pushed to db on wrong for loop.
I had already thought to check this, but the dbg messages were in the correct
for loop, and so output looked reasonable.
Little wonder I was having trouble parsing stuff; my RAM was loaded
with several times over with the db and the algo had to parse all of it.
Also fixed a segfault when 0.0.0.14 was randomly passed via random data
to the geolocation module.
Need to fix input so that doesn't happen.